Implement a cache for the eth_getBlockByHash call
This will cache block infos to avoid downloading the blocks again and
again for the gas price strategies.
